interplanetary-clock/backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pytest
pytest-mock
